AArch64: Improve A64FX memcpy
Hi Naohiro,

Since the A64FX memcpy is still quite large, I decided to do a quick pass to cleanup to
simplify the code. I believe the code is better overall and a bit faster, but please let me
know what you think. I've left the structure as it was, but there are likely more tweaks
possible. Here it is:

Reduce the codesize of the A64FX memcpy by avoiding duplication of code,
and removing redundant instructions. The size for memcpy and memmove
goes down from 1796 bytes to 1080 bytes. Performance is mostly unchanged
or slightly better as the critical loops are identical but fewer instructions are
executed before entering the loop.

Passes GLIBC regress, OK for commit?